refactor: extract Token enum to separate module
Move Token enum and its associated constants (KEYWORDS, SYMBOLS) to a dedicated token.rs module. This improves code organization by separating token definitions from lexer implementation.
