+++ compiler/src/lexer/testutils.rs
@@ -1,4 +1,4 @@
-use crate::lexer::{KEYWORDS, Name, SYMBOLS, Token};
+use crate::lexer::{Name, Token};
 
 use bolero::generator::{TypeGenerator, ValueGenerator as _, any, one_of, one_value_of};
 
@@ -10,8 +10,8 @@ const IDENTIFIERS: &[&str] = &[
 pub fn gen_token() -> impl bolero::ValueGenerator<Output = Token<'static>> {
     one_of((
         one_value_of(IDENTIFIERS).map_gen(|s| Token::Ident(Name::new(s))),
-        one_value_of(KEYWORDS).map_gen(|(_, t)| t),
-        one_value_of(SYMBOLS).map_gen(|(_, t)| t),
+        one_value_of(Token::KEYWORDS).map_gen(|(_, t)| t),
+        one_value_of(Token::SYMBOLS).map_gen(|(_, t)| t),
         any::<u64>().map_gen(Token::Num),
     ))
 }
