+++ compiler/src/lexer/token.rs
@@ -1,0 +1,85 @@
+use crate::common::Name;
+
+#[derive(Clone, Copy, Debug, PartialEq, Eq)]
+pub enum Token<'a> {
+    Fn,
+    Code,
+    Let,
+    Match,
+    LParen,
+    RParen,
+    LBracket,
+    RBracket,
+    LBrace,
+    RBrace,
+    Arrow,
+    Colon,
+    Eq,
+    Comma,
+    Semi,
+    Bar,
+    Ampersand,
+    EqEq,
+    Ne,
+    Lt,
+    Gt,
+    Le,
+    Ge,
+    Plus,
+    Minus,
+    Star,
+    Slash,
+    Bang,
+    HashLParen,
+    HashLBrace,
+    DollarLParen,
+    DollarLBrace,
+    DoubleLBracket,
+    DoubleRBracket,
+    DArrow,
+    Num(u64),
+    Ident(&'a Name),
+}
+
+impl Token<'static> {
+    pub const KEYWORDS: &[(&str, Self)] = &[
+        ("fn", Token::Fn),
+        ("code", Token::Code),
+        ("let", Token::Let),
+        ("match", Token::Match),
+    ];
+
+    pub const SYMBOLS: &[(&str, Self)] = &[
+        ("[[", Token::DoubleLBracket),
+        ("]]", Token::DoubleRBracket),
+        ("#(", Token::HashLParen),
+        ("#{", Token::HashLBrace),
+        ("$(", Token::DollarLParen),
+        ("${", Token::DollarLBrace),
+        ("=>", Token::DArrow),
+        (">=", Token::Ge),
+        ("<=", Token::Le),
+        ("==", Token::EqEq),
+        ("!=", Token::Ne),
+        ("->", Token::Arrow),
+        ("(", Token::LParen),
+        (")", Token::RParen),
+        ("[", Token::LBracket),
+        ("]", Token::RBracket),
+        ("{", Token::LBrace),
+        ("}", Token::RBrace),
+        (":", Token::Colon),
+        ("=", Token::Eq),
+        (",", Token::Comma),
+        (";", Token::Semi),
+        ("|", Token::Bar),
+        ("&", Token::Ampersand),
+        ("<", Token::Lt),
+        (">", Token::Gt),
+        ("+", Token::Plus),
+        ("-", Token::Minus),
+        ("*", Token::Star),
+        ("/", Token::Slash),
+        ("!", Token::Bang),
+    ];
+}
